EDGEAI-635 Changed image dims to use u32 instead of i32

+++ src/image.rs
@@ -247,12 +247,12 @@ impl Drop for ImageManager {
 #[derive(Debug)]
 pub struct Image {
     pub fd: OwnedFd,
-    pub width: i32,
-    pub height: i32,
+    pub width: u32,
+    pub height: u32,
     pub format: FourCC,
 }
 
-const fn format_row_stride(format: FourCC, width: i32) -> usize {
+const fn format_row_stride(format: FourCC, width: u32) -> usize {
     match format {
         RGB3 => 3 * width as usize,
         RGBX => 4 * width as usize,
@@ -263,12 +263,12 @@ const fn format_row_stride(format: FourCC, width: i32) -> usize {
     }
 }
 
-const fn image_size(width: i32, height: i32, format: FourCC) -> usize {
+const fn image_size(width: u32, height: u32, format: FourCC) -> usize {
     format_row_stride(format, width) * height as usize
 }
 
 impl Image {
-    pub fn new(width: i32, height: i32, format: FourCC) -> Result<Self, Box<dyn Error>> {
+    pub fn new(width: u32, height: u32, format: FourCC) -> Result<Self, Box<dyn Error>> {
         let heap = Heap::new(HeapKind::Cma)?;
         let fd = heap.allocate(image_size(width, height, format))?;
         Ok(Self {
@@ -279,7 +279,7 @@ impl Image {
         })
     }
 
-    pub fn new_preallocated(fd: OwnedFd, width: i32, height: i32, format: FourCC) -> Self {
+    pub fn new_preallocated(fd: OwnedFd, width: u32, height: u32, format: FourCC) -> Self {
         Self {
             fd,
             width,
@@ -300,11 +300,11 @@ impl Image {
         unsafe { DmaBuf::from_raw_fd(dup(self.fd.as_raw_fd())) }
     }
 
-    pub fn width(&self) -> i32 {
+    pub fn width(&self) -> u32 {
         self.width
     }
 
-    pub fn height(&self) -> i32 {
+    pub fn height(&self) -> u32 {
         self.height
     }
 
@@ -346,11 +346,11 @@ impl TryFrom<&Image> for g2d_surface {
             format: G2DFormat::from(img.format).format(),
             left: 0,
             top: 0,
-            right: img.width,
-            bottom: img.height,
-            stride: img.width,
-            width: img.width,
-            height: img.height,
+            right: img.width as i32,
+            bottom: img.height as i32,
+            stride: img.width as i32,
+            width: img.width as i32,
+            height: img.height as i32,
             blendfunc: 0,
             clrcolor: 0,
             rot: 0,
@@ -370,11 +370,11 @@ impl TryFrom<&Image> for g2d_surface_new {
             format: G2DFormat::from(img.format).format(),
             left: 0,
             top: 0,
-            right: img.width,
-            bottom: img.height,
-            stride: img.width,
-            width: img.width,
-            height: img.height,
+            right: img.width as i32,
+            bottom: img.height as i32,
+            stride: img.width as i32,
+            width: img.width as i32,
+            height: img.height as i32,
             blendfunc: 0,
             clrcolor: 0,
             rot: 0,
